Refactor: Move wall destruction logic into new system
